% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_bounds.R
\name{DD_bounds}
\alias{DD_bounds}
\title{Calculate the rejection interval's bounds}
\usage{
DD_bounds(
  metric,
  iterations = 15000,
  generate_plot = TRUE,
  alpha = 0.1,
  print_bounds = TRUE
)
}
\arguments{
\item{metric}{Vector containing metric values of each observation}

\item{iterations}{Number of iterations in generating the test distribution - the more, the better, but it is recommended to do at least 15'000 iterations, by default set to 15'000}

\item{generate_plot}{Boolean variable - set to TRUE if you want to show the density plot of test distribution, FALSE if you do not want to generate the plot; by default set to TRUE}

\item{alpha}{The significance level for calculating the rejection interval, by default set to 0.1}

\item{print_bounds}{Boolean variable - set to TRUE if you want to print the calculated bounds, FALSE if you don't want to print the bounds; by default set to TRUE}
}
\value{
Upper and lower bound of rejection interval, use 'DD_bounds(...)$lower' to print only the lower bound, 'DD_bounds(...)$upper' to print only the upper bound.
}
\description{
Using the metric's values, calculate the rejection interval, generate the test distribution's plot and print out the lower and upper bounds
}
