% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_e.R
\name{DD_e}
\alias{DD_e}
\title{Calculate the empirical Difference-in-Differences coefficient}
\usage{
DD_e(metric, time, affected)
}
\arguments{
\item{metric}{Vector containing metric values of each observation}

\item{time}{Boolean vector indicating whether an observation was taken pre-event or post-event - FALSE (or 0) if the observation was taken before the event took place, TRUE (or 1) if the observation was taken after the event took place.}

\item{affected}{Boolean vector indicating whether an observation was to be affected by the event - FALSE (or 0) for control group, TRUE (or 1) for experimental (affected) group.}
}
\value{
The numerical value of Difference-in-Differences coefficient
}
\description{
Given metric values, a Boolean vector indicating whether an observation belongs to control or experimental group and another Boolean vector indicating whether an observation was taken pre-event or post-event, calculate the empirical DD coefficient.
}
