% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_gt_table.R
\name{DD_gt_table}
\alias{DD_gt_table}
\title{Generate the differences table as a 'gt' object}
\usage{
DD_gt_table(
  metric,
  time,
  affected,
  table_title = "DD Estimation",
  post_event_name = "post-event",
  pre_event_name = "pre-event",
  affected_group_name = "Experimental",
  control_group_name = "Control",
  time_header = "Time",
  affected_header = "Affected"
)
}
\arguments{
\item{metric}{Vector containing metric values of each observation}

\item{time}{Boolean vector indicating whether an observation was taken pre-event or post-event - FALSE (or 0) if the observation was taken before the event took place, TRUE (or 1) if the observation was taken after the event took place.}

\item{affected}{Boolean vector indicating whether an observation was to be affected by the event - FALSE (or 0) for control group, TRUE (or 1) for experimental (affected) group.}

\item{table_title}{The title for the whole table, 'DD Estimation' by default}

\item{post_event_name}{Name for post-event subgroup in the differences table, set to 'post-event' by default}

\item{pre_event_name}{Name for pre-event subgroup in the differences table, set to 'pre-event' by default}

\item{affected_group_name}{Name for affected subgroup in the differences table, set to 'experimental' by default}

\item{control_group_name}{Name for control subgroup in the differences table, set to 'control' by default}

\item{time_header}{Header name for the variable describing whether an observation was taken pre-event or post-event, set to 'Time' as default}

\item{affected_header}{Header name for the variable describing whether an observation belongs to control group or affected group, set to 'Affected' as default}
}
\value{
'gt' object containing publication-ready visualization of DD calculations using differences table
}
\description{
Generate the visualization of calculation process using differences table with customized headings and group names
}
