% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktd_DD_bounds.R
\name{ktd_DD_bounds}
\alias{ktd_DD_bounds}
\title{Calculate the rejection interval's bounds on known test distribution}
\usage{
ktd_DD_bounds(
  test_data,
  generate_plot = TRUE,
  alpha = 0.1,
  print_bounds = TRUE
)
}
\arguments{
\item{test_data}{Previously generated test distribution, can be generated using 'test_distr()' function}

\item{generate_plot}{Boolean variable - set to TRUE if you want to show the density plot of test distribution, FALSE if you do not want to generate the plot; by default set to TRUE}

\item{alpha}{The significance level for calculating the rejection interval, by default set to 0.1}

\item{print_bounds}{Boolean variable - set to TRUE if you want to print the calculated bounds, FALSE if you don't want to print the bounds; by default set to TRUE}
}
\value{
Upper and lower bound of rejection interval, use 'DD_bounds(...)$lower' to print only the lower bound, 'DD_bounds(...)$upper' to print only the upper bound.
}
\description{
Using the previously generated test distribution, calculate the rejection interval, generate the test distribution's plot and print out the lower and upper bounds
}
