% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktd_DD_sig.R
\name{ktd_DD_sig}
\alias{ktd_DD_sig}
\title{Determine whether the empirical DD coefficient is significant}
\usage{
ktd_DD_sig(
  test_data,
  DDe,
  alpha = 0.1,
  generate_plot = TRUE,
  print_bounds = TRUE
)
}
\arguments{
\item{test_data}{Previously generated test distribution, can be generated using 'test_distr()' function}

\item{DDe}{The empirical value of DD coefficient, can be calculated by DDe() function}

\item{alpha}{The significance level for calculating the rejection interval, by default set to 0.1}

\item{generate_plot}{Boolean variable - set to TRUE if you want to show the density plot of test distribution, FALSE if you do not want to generate the plot; by default set to TRUE}

\item{print_bounds}{Boolean variable - set to TRUE if you want to print the calculated bounds, FALSE if you don't want to print the bounds; by default set to TRUE}
}
\description{
Using the simulated method, calculate the p-value and determine whether empirical DD coefficient is statistically different than 0
}
