% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_DD.R
\name{table_to_DD}
\alias{table_to_DD}
\title{Convert Table to Data Frame suitable for DD modelling}
\usage{
table_to_DD(
  period,
  seriesA,
  seriesB,
  dateX = median(1:floor((length(period))))
)
}
\arguments{
\item{period}{A numeric vector representing dates or periods. Can be either
in a date format or a simple numeric vector, with consecutive numbers representing
consecutive periods}

\item{seriesA}{A numeric vector representing values for series A.}

\item{seriesB}{A numeric vector representing values for series B.}

\item{dateX}{An optional numeric value representing the reference date 
for splitting the data into "before" and "after" categories. Defaults 
to the median of the `date` vector.}
}
\value{
A data frame with three columns:
  - metric: A combined vector containing values from both seriesA and seriesB.
  - affected: A factor indicating series (1 for seriesA, 0 for seriesB).
  - time: A factor indicating time period (0 for before `dateX`, 1 for after `dateX`).
}
\description{
This function takes three numeric vectors representing date, seriesA, 
and seriesB values and converts them into a data frame suitable for modelling
with sigDD package
}
