% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_trends_plot.R
\name{parallel_trends_plot}
\alias{parallel_trends_plot}
\title{Parallel Trend Plot}
\usage{
parallel_trends_plot(
  metric,
  period,
  affected,
  dateX = ceiling(median(time)),
  is_date = FALSE,
  control_mod = 0,
  affected_mod = 0,
  equalize = FALSE,
  unitize_period = FALSE,
  title = "Parallel Trend Plot",
  xlabel = "Time",
  ylabel = "Metric",
  control_series_color = "red",
  affected_series_color = "blue",
  legend_color_label = "Affected",
  control_name = "0",
  affected_name = "1"
)
}
\arguments{
\item{metric}{Vector containing metric values of each observation}

\item{period}{A vector of dates (or a number sequence) representing the period (f.e. day) when the observation was taken}

\item{affected}{Boolean vector indicating whether an observation was to be affected by the event - FALSE (or 0) for control group, TRUE (or 1) for experimental (affected) group.}

\item{dateX}{An optional numeric value representing the time point for
the vertical line. Defaults to the median of `time`.}

\item{is_date}{A logical value (TRUE/FALSE) indicating if `time` is a date
vector. If TRUE, `time` should be a character vector in a recognized date
format. The starting date is assumed to be the first element.}

\item{control_mod}{A numeric value to modify the metric for the control group.
Defaults to 0 (no modification).}

\item{affected_mod}{A numeric value to modify the metric for the affected group.
Defaults to 0 (no modification).}

\item{equalize}{A logical value (TRUE/FALSE) indicating whether to equalize
the mean metric values of the groups.}

\item{unitize_period}{A logical value (TRUE/FALSE) indicating whether to unitize
the time axis (scale to range 0-1).}

\item{title}{The main title of the plot. Defaults to "Parallel Trend Plot".}

\item{xlabel}{The label for the x-axis. Defaults to "Time".}

\item{ylabel}{The label for the y-axis. Defaults to "Metric".}

\item{control_series_color}{The color for the control group line. Defaults to "red".}

\item{affected_series_color}{The color for the affected group line. Defaults to "blue".}

\item{legend_color_label}{The label for the color legend. Defaults to "Affected".}

\item{control_name}{The name for the control group, '0' by default}

\item{affected_name}{The name for the affected group, '1' by default}
}
\value{
A ggplot object representing the parallel trend plot.
}
\description{
Create a parallel trend plot to visualize trends
between groups over time.
}
