% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_simple_test.R
\name{rg_simple_test}
\alias{rg_simple_test}
\title{Perform a Simple Parallel Trend Test}
\usage{
rg_simple_test(
  period,
  dateX,
  metric,
  time,
  affected,
  alpha = 0.1,
  additional_variable_matrix = matrix(nrow = 1, ncol = 1)
)
}
\arguments{
\item{period}{A numeric vector representing the period (time point) for 
each observation.}

\item{dateX}{A numeric value representing the reference date (time point) 
for splitting the data into "pre-treatment" and "post-treatment" periods.}

\item{metric}{Vector containing metric values of each observation}

\item{time}{Boolean vector indicating whether an observation was taken pre-event or post-event - FALSE (or 0) if the observation was taken before the event took place, TRUE (or 1) if the observation was taken after the event took place.}

\item{affected}{Boolean vector indicating whether an observation was to be affected by the event - FALSE (or 0) for control group, TRUE (or 1) for experimental (affected) group.}

\item{alpha}{A numeric value (between 0 and 1) representing the significance 
level for the test. Defaults to 0.1.}

\item{additional_variable_matrix}{An optional matrix (default is a 1x1 
empty matrix) allowing for inclusion of additional independent variables 
in the test model. The number of rows must either be 1 (empty matrix) or 
equal to the length of the `metric` vector.}
}
\value{
A data frame with two columns:
  - t test: The t-statistic value from the linear model.
  - p-value: The p-value associated with the t-statistic.
}
\description{
This function performs a simple statistical test to assess whether 
parallel trends exist between groups over time.
}
